Oculus now works with firstpersoncontrols.js
Oculus is working. Plan on finding a way to use our PointerLockControls
with the oculus so you can use it to look around while you move with
the wii balance board. This is a place holder for the time being. Now
just need to make a nice way of switching between the oculus view and
the standard view.